Feat: 사장님-가게정보상세-타입스크립트 변환
